+++ lib/presentation/widgets/attachment/desktop_attachment.dart
@@ -76,7 +76,7 @@ class _DesktopAttachmentState extends State<DesktopAttachment> {
         scrollDirection: Axis.vertical,
         child: Column(
           children: [
-            const TopHeaders(),
+            const TopHeaders(isDesktop: true),
             Row(
               children: [
                 Expanded(